mcp_example/pyproject.toml
[project]
name = "mcp-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-api-python-client>=2.170.0",
    "google-auth>=2.40.2",
    "httpx>=0.28.1",
    "langchain-core>=0.3.61",
    "langchain-mcp-adapters>=0.1.1",
    "langchain-openai>=0.3.18",
    "langgraph>=0.4.7",
    "mcp[cli]>=1.9.1",
    "openai>=1.82.0",
    "psutil>=7.0.0",
    "python-dotenv>=1.1.0",
]
